LPS-65012 DDL move constant class back and export the package
